# openapi/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Go Currency Tracker API",
    "description": "API for getting currency rates from the Central Bank of Russia with web interface for analysis",
    "version": "1.2.0",
    "contact": {
      "name": "Go Currency Tracker Team"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development server"
    }
  ],
  "paths": {
    "/": {
      "get": {
        "summary": "Web interface",
        "description": "Web interface for currency rate analysis",
        "operationId": "webInterface",
        "responses": {
          "200": {
            "description": "HTML page with web interface",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/ping": {
      "get": {
        "summary": "API health check",
        "description": "Used to check API availability",
        "operationId": "ping",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string",
                  "example": "pong"
                }
              }
            }
          }
        }
      }
    },
    "/info": {
      "get": {
        "summary": "Service information",
        "description": "Returns basic information about the service",
        "operationId": "getInfo",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "service": {
                      "type": "string",
                      "example": "Go Currency Tracker"
                    },
                    "status": {
                      "type": "string",
                      "example": "OK"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/rates/cbr": {
      "get": {
        "summary": "Get all currency rates from CBR",
        "description": "Returns all currency rates from the Central Bank of Russia for the specified date",
        "operationId": "getCBRRates",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "description": "Date in YYYY-MM-DD format (e.g., 2023-05-15). If not specified, current date is used.",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date",
              "example": "2023-05-15"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "$ref": "#/components/schemas/Valute"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string",
                  "example": "Error getting rates"
                }
              }
            }
          }
        }
      }
    },
    "/rates/cbr/currency": {
      "get": {
        "summary": "Get specific currency rate from CBR",
        "description": "Returns the rate of the specified currency from the Central Bank of Russia for the specified date",
        "operationId": "getCBRCurrency",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "description": "Currency code in ISO 4217 format (e.g., USD, EUR)",
            "required": true,
            "schema": {
              "type": "string",
              "example": "USD"
            }
          },
          {
            "name": "date",
            "in": "query",
            "description": "Date in YYYY-MM-DD format (e.g., 2023-05-15). If not specified, current date is used.",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date",
              "example": "2023-05-15"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Valute"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string",
                  "example": "Currency code not specified"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string",
                  "example": "Error getting currency rate"
                }
              }
            }
          }
        }
      }
    },
    "/rates/cbr/history": {
      "get": {
        "summary": "Get historical currency rates",
        "description": "Returns historical rates for the specified currency over a specified number of days",
        "operationId": "getCurrencyHistory",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "description": "Currency code in ISO 4217 format (e.g., USD, EUR)",
            "required": true,
            "schema": {
              "type": "string",
              "example": "USD"
            }
          },
          {
            "name": "days",
            "in": "query",
            "description": "Number of days to look back (max 365)",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 365,
              "example": 30
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "date": {
                            "type": "string",
                            "format": "date",
                            "example": "2023-05-15"
                          },
                          "code": {
                            "type": "string",
                            "example": "USD"
                          },
                          "name": {
                            "type": "string",
                            "example": "US Dollar"
                          },
                          "nominal": {
                            "type": "integer",
                            "example": 1
                          },
                          "value": {
                            "type": "number",
                            "format": "float",
                            "example": 75.4571
                          },
                          "previous": {
                            "type": "number",
                            "format": "float",
                            "example": 75.2513
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "error": {
                      "type": "string",
                      "example": "Currency code not specified"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "error": {
                      "type": "string",
                      "example": "Error getting historical data"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Valute": {
        "type": "object",
        "properties": {
          "ID": {
            "type": "string",
            "description": "Internal currency identifier in CBR API",
            "example": "R01235"
          },
          "NumCode": {
            "type": "string",
            "description": "Numeric currency code",
            "example": "840"
          },
          "CharCode": {
            "type": "string",
            "description": "Alphabetic currency code according to ISO 4217",
            "example": "USD"
          },
          "Nominal": {
            "type": "integer",
            "description": "Nominal (number of currency units)",
            "example": 1
          },
          "Name": {
            "type": "string",
            "description": "Currency name",
            "example": "US Dollar"
          },
          "Value": {
            "type": "number",
            "format": "float",
            "description": "Currency rate to Russian ruble",
            "example": 75.4571
          },
          "Previous": {
            "type": "number",
            "format": "float",
            "description": "Previous currency rate value",
            "example": 75.2513
          }
        }
      }
    }
  }
}